layout/style/test/test_bug373293.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=373293
-->
<head>
  <title>Test for Bug 373293</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=373293">Mozilla Bug 373293</a>
<p id="display"></p>
<div id="content" style="display: none">
  <div id="t" style="color: transparent;"></div>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 373293**/

var actual = $("t").getAttribute("style");
var expected = "color: transparent;";
is(actual, expected, "Expected style content did not match the actual style content");

</script>
</pre>
</body>
</html>
